More singly linked lists

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,14 +1,15 @@
 #include "lists.h"
 
 /**
-* listint_len - function that returns the number of elements in a linked listint_t list.
+* listint_len - returns the number of elements in a linked listint_t list.
 * @h: pointer
 * Return: the number of elements
 */
 
 size_t listint_len(const listint_t *h)
 {
-	int i = 0;
+	size_t i = 0;
+
 	k = h;
 	while (k != NULL)
 	{
